Add profile to resolve self-referential dependency
The 'built' profile will add the remapped dependency generated
in the initialize phase. This profile is on by default since
during normal usage the mod will have it as a dependency. H

However, before the artifact is installed, it cannot reliably
be used as a dependency. To fix this the profile can be negated,
temporarily removing the dependency so it can be installed:

mvn initialize -P -built
mvn package