Clear ACC_SYNTHETIC flag on all symbols
Fixes mod compilation not finding required symbols in the
remapped Minecraft dependency, for example: if your mod
subclasses CommandBase, which implements ICommand and in turn
implements Comparable, but fails to compile with an error
compareTo() unimplemented.

compareTo(Object) is in fact implemented in CommandBase,
and in turn calls compareNameTo(String). However, compareTo()
is flagged as ACC_SYNTHETIC, meaning it was auto-generated by
the compiler as a bridge method and is not present in the
original source code. This hides it from being referenced when
compiled against as a library, for type safety purposes.

This problem does not occur when using MCP's decompile + recompile
process; fernflower comments '// $FF: synthetic method' (full list: https://gist.github.com/agaricusb/5122279)
but recompilation destroys the synthetic flag, since by definition,
the method is no longer auto-generated and is present in the source.

Therefore, for compatibility with mods written for MCP's process,
we clear this synthetic bit to expose all symbols everywhere.

Thanks to cpw and LexManos for helping diagnose this issue.